fix: config uses floats

+++ include/config.hpp
@@ -4,7 +4,7 @@ inline constexpr char TITLE[]{"Abhay's Path Tracer"};
 
 inline constexpr auto WIDTH{1280};
 inline constexpr auto HEIGHT{720};
-inline constexpr auto ASPECT{static_cast<double>(WIDTH) / HEIGHT};
+inline constexpr auto ASPECT{static_cast<float>(WIDTH) / HEIGHT};
 
-inline constexpr auto FOV{90.0};
+inline constexpr auto FOV{90.0f};
 inline constexpr auto SAMPLE_COUNT{10};
